Finished
So on desktop it's going to be a horizontal tab, but on mobile or screens less than 1100px we have a drop down. The indicated/active tab persists even when resizing. As well we have the image carousel working with the buttons working too. The main issues were structuring how it would look in the code, preventing duplicate event listeners, getting active tab to persist, etc. It was more annoying than necessary, but still I understand why now.